usertests

+++ routes/userRouter.js
@@ -1,25 +1,52 @@
 const express = require('express')
 const bcrypt = require('bcrypt')
 const User = require('../models/user')
+const config = require('../utils/config')
 const router = express.Router()
+
+const validateEmail = async (addr) => {
+  if (!addr) return null
+  if (
+    /^[-!#$%&'*+/0-9=?A-Z^_a-z{|}~](.?[-!#$%&'*+/0-9=?A-Z^_a-z`{|}~])*@[a-zA-Z0-9](-*.?[a-zA-Z0-9])*.[a-zA-Z](-?[a-zA-Z0-9])+$/.test(
+      addr
+    )
+  ) {
+    if (await User.findOne({ email: addr }))
+      throw {
+        name: 'Custom',
+        message: 'email already in use',
+      }
+    return addr
+  }
+  throw {
+    name: 'Custom',
+    message: 'invalid email',
+  }
+}
 
 router.post('/', async (req, res) => {
   const data = req.body
-  if (data['password'].length < 5 || data['password'].length > 100) {
+  if (!data.username || !data.password)
     throw {
       name: 'Custom',
-      message: 'password length must be 5-100 characters',
+      message: 'provide username and password',
     }
-  }
+  if (data['password'].length < 5 || data['password'].length > 50)
+    throw {
+      name: 'Custom',
+      message: 'password length must be 5-50 characters',
+    }
+
   const hash = await bcrypt.hash(data['password'], 10)
+
   const newUser = User({
     username: data['username'],
     passwordhash: hash,
-    email: data['email'],
+    email: await validateEmail(data['email']),
   })
   const response = await newUser.save()
 
-  res.json(response).status(201).end()
+  res.status(201).json(response).end()
 })
 
 router.get('/', async (req, res) => {
@@ -41,20 +68,18 @@ router.get('/', async (req, res) => {
         model: 'Sport',
       },
     })
-
     .populate('games')
   res.json(users).status(200).end()
 })
 
 router.delete('/:id', async (req, res) => {
   await User.findByIdAndDelete(req.params.id)
-
   res.status(204).end()
 })
 
 router.delete('/', async (req, res) => {
+  if (config.NODE_ENV !== 'test') return
   await User.deleteMany({})
-
   res.status(204).end()
 })
